src/pyproject.toml
[project]
name = "symbology"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "edgartools==3.15.0",
  "fastapi==0.115.12",
  "httpx==0.28.1",
  "ollama==0.4.7",
  "tiktoken",
  "transformers",
  "torch",
  "psycopg2==2.9.10",
  "pydantic==2.11.3",
  "pydantic-settings==2.9.0",
  "pytest>=8.3.5",
  "pytest-cov>=6.1.1",
  "pytest-json-report>=1.5.0",
  "python-dotenv==1.0.0",
  "requests>=2.32.3",
  "safe-result",
  "sqlalchemy==2.0.40",
  "structlog==24.1.0",
  "uuid7==0.1.0",
  "uvicorn==0.34.1",
  "matplotlib==3.10.3",
  "seaborn==0.13.2",
  "jupyter",
  'ipywidgets',
]

# [tool.uv.workspace]
# members = ["src/python"]
[tool.pytest.ini_options]
filterwarnings = [
  "ignore::DeprecationWarning:pydantic.*:",
  "ignore::UserWarning:pydantic.*:",
]

[tool.ruff]
# Enable flake8-bugbear (`B`) rules, and isort (`I`) for import sorting
lint.select = ["E", "F", "B", "I"]
# Allow unused variables when underscore-prefixed
lint.dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"
# Exclude a variety of commonly ignored directories
exclude = [".git", ".eggs", "__pycache__", "build", "dist", "*.ipynb"]
# Same as Black
line-length = 160
# Target Python 3.13
target-version = "py313"

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"] # Ignore unused imports in __init__ files
"**llm*" = ["E501"]      # Ignore long lines in prompts, etc
"conftest.py" = ["F401"]

[tool.ruff.lint.isort]
known-first-party = ["symbology"]
# Enable automatic import sorting when running Ruff
# automatically-sort-imports = true
# Ensure imports are properly organized
force-single-line = false
force-sort-within-sections = true
order-by-type = false
